change some qrcode method

+++ mikoto/src/com/s1025/mikoto/api/QRCodeApi.java
@@ -6,14 +6,14 @@ import com.s1025.mikoto.util.HttpCon;
 public class QRCodeApi {
 	public String create_qrcode_url = "https://api.weixin.qq.com/cgi-bin/qrcode/create?access_token=ACCESS_TOKEN";
 	
-	public String createQRCode(int seconds, int scene){
+	public String createQRCode(int seconds, long scene){
 		String url = create_qrcode_url.replace("ACCESS_TOKEN", Mikoto.api.access.getAccessToken());
 		String post = "{\"expire_seconds\": "+seconds+", \"action_name\": \"QR_SCENE\", \"action_info\": {\"scene\": {\"scene_id\": "+scene+"}}}";
 		String response = HttpCon.httpRequest(url, "POST", post);
 		return response;
 	}
 	
-	public String createLimitQRCode(int scene){
+	public String createLimitQRCode(long scene){
 		String url = create_qrcode_url.replace("ACCESS_TOKEN", Mikoto.api.access.getAccessToken());
 		String post = "{\"action_name\": \"QR_LIMIT_SCENE\", \"action_info\": {\"scene\": {\"scene_id\": "+scene+"}}}";
 		String response = HttpCon.httpRequest(url, "POST", post);
